Fix bugs for tests
